+++ packages/components/src/navigator/index.ts
@@ -1,5 +1,5 @@
 export { default as NavigatorProvider } from './navigator-provider';
 export { default as NavigatorScreen } from './navigator-screen';
-export { default as NavigatorLink } from './navigator-link';
-export { default as NavigatorBackLink } from './navigator-back-link';
+export { default as NavigatorButton } from './navigator-button';
+export { default as NavigatorBackButton } from './navigator-back-button';
 export { default as useNavigator } from './use-navigator';