+++ packages/components/src/navigator/navigator-button/component.tsx
@@ -9,51 +9,54 @@ import type { Ref } from 'react';
  */
 import { contextConnect, WordPressComponentProps } from '../../ui/context';
 import { View } from '../../view';
-import { useNavigatorLink } from './hook';
-import type { NavigatorLinkProps } from '../types';
+import { useNavigatorButton } from './hook';
+import type { NavigatorButtonProps } from '../types';
 
-function NavigatorLink(
-	props: WordPressComponentProps< NavigatorLinkProps, 'button' >,
+function NavigatorButton(
+	props: WordPressComponentProps< NavigatorButtonProps, 'button' >,
 	forwardedRef: Ref< any >
 ) {
-	const navigatorLinkProps = useNavigatorLink( props );
+	const navigatorButtonProps = useNavigatorButton( props );
 
-	return <View ref={ forwardedRef } { ...navigatorLinkProps } />;
+	return <View ref={ forwardedRef } { ...navigatorButtonProps } />;
 }
 
 /**
- * The `NavigatorLink` component can be used to navigate to a screen and should
+ * The `NavigatorButton` component can be used to navigate to a screen and should
  * be used in combination with the `NavigatorProvider`, the `NavigatorScreen`
- * and the `NavigatorBackLink` components (or the `useNavigator` hook).
+ * and the `NavigatorBackButton` components (or the `useNavigator` hook).
  *
  * @example
  * ```jsx
  * import {
  *   __experimentalNavigatorProvider as NavigatorProvider,
  *   __experimentalNavigatorScreen as NavigatorScreen,
- *   __experimentalNavigatorLink as NavigatorLink,
- *   __experimentalNavigatorBackLink as NavigatorBackLink,
+ *   __experimentalNavigatorButton as NavigatorButton,
+ *   __experimentalNavigatorBackButton as NavigatorBackButton,
  * } from '@wordpress/components';
  *
  * const MyNavigation = () => (
  *   <NavigatorProvider initialPath="/">
  *     <NavigatorScreen path="/">
  *       <p>This is the home screen.</p>
- *        <NavigatorLink path="/child">
+ *        <NavigatorButton path="/child">
  *          Navigate to child screen.
- *       </NavigatorLink>
+ *       </NavigatorButton>
  *     </NavigatorScreen>
  *
  *     <NavigatorScreen path="/child">
  *       <p>This is the child screen.</p>
- *       <NavigatorBackLink>
+ *       <NavigatorBackButton>
  *         Go back
- *       </NavigatorBackLink>
+ *       </NavigatorBackButton>
  *     </NavigatorScreen>
  *   </NavigatorProvider>
  * );
  * ```
  */
-const ConnectedNavigatorLink = contextConnect( NavigatorLink, 'NavigatorLink' );
+const ConnectedNavigatorButton = contextConnect(
+	NavigatorButton,
+	'NavigatorButton'
+);
 
-export default ConnectedNavigatorLink;
+export default ConnectedNavigatorButton;